requirements.txt
Django<=5.0
djangorestframework==3.14.0
django-celery-beat==2.5.0
Celery==5.3.6
django-environ==0.11.2
kafka-python==2.0.2
